requirements.txt
langchain==0.2.5
chromadb==0.4.24
faiss-cpu==1.7.4
openai==1.35.14
tiktoken==0.7.0
sentence-transformers==2.6.1
pypdf==4.2.0
unstructured==0.14.7
pdfplumber==0.11.0
streamlit==1.36.0
python-dotenv==1.0.1
numpy==1.26.4
pandas==2.2.2
